requirements.txt
boto3==1.28.64
moto==4.1.3
pandas==2.0.3
python_dateutil==2.8.2
SQLAlchemy==1.4.52
